+++ ui/quasar.config.ts
@@ -12,12 +12,7 @@ export default defineConfig((/* ctx */) => {
     // app boot file (/src/boot)
     // --> boot files are part of "main.js"
     // https://v2.quasar.dev/quasar-cli-vite/boot-files
-    boot: [
-      'axios',
-      'code-block',
-      // code editor
-      'monaco-editor',
-    ],
+    boot: ['axios', 'monaco-editor'],
 
     // https://v2.quasar.dev/quasar-cli-vite/quasar-config-file#css
     css: ['app.scss'],