+++ ui/src/constant/data/query-str-example.ts
@@ -1,0 +1,8 @@
+/**
+ *
+ * @author shiloh
+ * @date 2025/3/25 22:00
+ */
+
+// URL Query String to test
+export const TEST_QUERY_STR = 'name=John&age=30&city=New York&hobbies=coding&hobbies=cooking'